test redo for cards complete
